+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## X.Y.Z 2022-xx-xx
+* [Fixed] Fixed a bug where 3 decimal place currencies were not being formatted properly.
+
 ### PaymentSheet
 * [Fixed] Fixed an issue that caused animations of the card logos in the Card input field to glitch.
 * [Fixed] Fixed a layout issue in the "Save my info" checkbox.